+++ cmp/treemap/SplitTreeMap.ts
@@ -4,16 +4,14 @@
  *
  * Copyright © 2024 Extremely Heavy Industries Inc.
  */
-import {AgGrid} from '@xh/hoist/cmp/ag-grid';
-import {box, div, fragment, hbox, hframe, p, placeholder, vbox, vframe} from '@xh/hoist/cmp/layout';
+import {box, div, fragment, hbox, hframe, placeholder, vbox, vframe} from '@xh/hoist/cmp/layout';
 import {BoxProps, hoistCmp, HoistProps, uses, XH} from '@xh/hoist/core';
-import {errorMessage} from '@xh/hoist/desktop/cmp/error';
-import {mask} from '@xh/hoist/desktop/cmp/mask';
-import '@xh/hoist/desktop/register';
-import classNames from 'classnames';
+import {splitTreeMapImpl as desktopSplitTreeMapImpl} from '@xh/hoist/dynamics/desktop';
+import {splitTreeMapImpl as mobileSplitTreeMapImpl} from '@xh/hoist/dynamics/mobile';
+import {AgGrid} from '@xh/hoist/cmp/ag-grid';
 import {compact, uniq} from 'lodash';
-import {splitter} from './impl/Splitter';
-
+import {ReactNode} from 'react';
+import classNames from 'classnames';
 import './SplitTreeMap.scss';
 import {SplitTreeMapModel} from './SplitTreeMapModel';
 import {treeMap} from './TreeMap';
@@ -27,33 +25,27 @@ export const [SplitTreeMap, splitTreeMap] = hoistCmp.withFactory<SplitTreeMapPro
     displayName: 'SplitTreeMap',
     model: uses(SplitTreeMapModel),
     className: 'xh-split-treemap',
-
     render({model, className, testId, ...props}, ref) {
-        const {primaryMapModel, secondaryMapModel, orientation} = model,
-            errors = uniq(compact([primaryMapModel.error, secondaryMapModel.error])),
-            container = orientation === 'horizontal' ? hframe : vframe;
+        const container = model.orientation === 'horizontal' ? hframe : vframe,
+            childMaps = getChildMaps(model);
 
         return container({
             testId,
             ref,
             className,
-            items: errors.length ? errorPanel({errors}) : childMaps(),
+            item: XH.isMobileApp
+                ? mobileSplitTreeMapImpl(model, childMaps)
+                : desktopSplitTreeMapImpl(model, childMaps),
             ...props
         });
     }
 });
 
-const childMaps = hoistCmp.factory<SplitTreeMapModel>(({model}) => {
-    const {
-        primaryMapModel,
-        secondaryMapModel,
-        empty,
-        emptyText,
-        isMasking,
-        mapTitleFn,
-        showSplitter
-    } = model;
-    if (empty) return placeholder(emptyText);
+function getChildMaps(model: SplitTreeMapModel): ReactNode {
+    const {primaryMapModel, secondaryMapModel, empty, emptyText, mapTitleFn, showSplitter} = model,
+        errors = uniq(compact([primaryMapModel.error, secondaryMapModel.error]));
+
+    if (empty || errors.length) return placeholder(emptyText);
 
     const pTotal = primaryMapModel.total,
         sTotal = secondaryMapModel.total;
@@ -89,16 +81,9 @@ const childMaps = hoistCmp.factory<SplitTreeMapModel>(({model}) => {
             flex: sFlex,
             className: 'xh-split-treemap__map-holder',
             item: treeMap({model: secondaryMapModel})
-        }),
-
-        // Mask
-        div({
-            omit: !isMasking,
-            className: 'xh-split-treemap__mask-holder',
-            item: mask({isDisplayed: true, spinner: true})
         })
     ]);
-});
+}
 
 const mapTitle = hoistCmp.factory<SplitTreeMapModel>(({model, isPrimary}) => {
     const {mapTitleFn, orientation} = model,
@@ -110,7 +95,8 @@ const mapTitle = hoistCmp.factory<SplitTreeMapModel>(({model, isPrimary}) => {
 
     const container = titleOrientation === 'horizontal' ? hbox : vbox,
         dim = titleOrientation === 'horizontal' ? 'height' : 'width',
-        size = (AgGrid as any).getHeaderHeightForSizingMode(XH.sizingMode);
+        sizingMode = XH.isMobileApp ? 'tiny' : XH.sizingMode,
+        size = (AgGrid as any).getHeaderHeightForSizingMode(sizingMode);
 
     return container({
         style: {[dim]: `${size}px`},
@@ -125,6 +111,12 @@ const mapTitle = hoistCmp.factory<SplitTreeMapModel>(({model, isPrimary}) => {
     });
 });
 
-const errorPanel = hoistCmp.factory(({errors}) =>
-    errorMessage({error: errors.join(' '), message: fragment(errors.map(e => p(e)))})
-);
+const splitter = hoistCmp.factory<SplitTreeMapModel>(({model}) => {
+    const {orientation} = model,
+        vertical = orientation === 'vertical',
+        cmp = vertical ? hbox : vbox;
+
+    return cmp({
+        className: `xh-split-treemap__splitter xh-split-treemap__splitter--${vertical ? 'vertical' : 'horizontal'}`
+    });
+});